requirements.txt
requests>=2.31.0
dnspython>=2.4.2
cryptography>=41.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
python-whois>=0.8.0
